Fixed issue where maximum memory for capture couldn't be set higher than 2 GiB due to use of 32bit int in calculations.  uint64_t is now used to support larger buffer sizes.

+++ native/src/ProfilerPlugin.cpp
@@ -526,7 +526,7 @@ extern "C" int UNITY_INTERFACE_EXPORT UNITY_INTERFACE_API BeginCapture(const cha
 		return 0;
 	}
 
-	gCaptureState.maxMemoryUsageBytes = maxMemUsageMB * 1024 * 1024; // MB to B
+	gCaptureState.maxMemoryUsageBytes = static_cast<uint64_t>(maxMemUsageMB) * 1024 * 1024; // MB to B
 	gCaptureState.captureStartTime = std::chrono::steady_clock::now();
 	gCaptureState.filename = filename;
 	gCaptureState.inFrame = false;
